fix: price Long -> Double 대대적 변경

+++ backend/funco-rank/src/main/java/com/found_404/funco/feignClient/service/TradeService.java
@@ -22,7 +22,7 @@ public class TradeService {
 
 	private final TradeServiceClient tradeServiceClient;
 
-	public Map<String, Long> getCryptoPrice(List<String> tickers) {
+	public Map<String, Double> getCryptoPrice(List<String> tickers) {
 		try {
 			return tradeServiceClient.getCryptoPrice(tickers);
 		} catch (
